Network: Move training utilities into separate package
Also refactor existing scripts using them.
